+++ src/xml/xml_native_reader.h
@@ -101,7 +101,7 @@ class mjXReader : public mjXBase {
 };
 
 // MJCF schema
-#define nMJCF 239
+#define nMJCF 237
 extern const char* MJCF[nMJCF][mjXATTRNUM];
 
 #endif  // MUJOCO_SRC_XML_XML_NATIVE_READER_H_